use generic component in elvlButtonGroup
