+++ packages/client/src/components/index.tsx
@@ -1,9 +1,15 @@
+import { AttributeIcon } from "./basic/AttributeIcon/AttributeIcon";
 import { Box } from "./basic/Box/Box";
 import { Button } from "./basic/Button/Button";
 import { ButtonGroup, ButtonGroups } from "./basic/ButtonGroup/ButtonGroup";
+import { Checkbox } from "./basic/Checkbox/Checkbox";
+import { Cloud } from "./basic/Cloud/Cloud";
 import { Dropdown } from "./basic/Dropdown/Dropdown";
+import { Dropzone } from "./basic/Dropzone/Dropzone";
 import { Header } from "./basic/Header/Header";
+import { IconButtonGroup } from "./basic/IconButtonGroup/IconButtonGroup";
 import { Input } from "./basic/Input/Input";
+import { LetterIcon } from "./basic/LetterIcon/LetterIcon";
 import { Loader } from "./basic/Loader/Loader";
 import {
   Modal,
@@ -19,19 +25,13 @@ import { MultiInput } from "./basic/MultiInput/MultiInput";
 import { Panel } from "./basic/Panel/Panel";
 import { Submit } from "./basic/Submit/Submit";
 import { Suggester } from "./basic/Suggester/Suggester";
+import { Table } from "./basic/Table/Table";
 import { Tag } from "./basic/Tag/Tag";
 import { TagGroup } from "./basic/TagGroup/TagGroup";
-import { Toast } from "./basic/Toast/Toast";
-import { AttributeIcon } from "./basic/AttributeIcon/AttributeIcon";
-import { Checkbox } from "./basic/Checkbox/Checkbox";
-import { Table } from "./basic/Table/Table";
-import { TypeBar } from "./basic/TypeBar/TypeBar";
-import { Cloud } from "./basic/Cloud/Cloud";
-import { LetterIcon } from "./basic/LetterIcon/LetterIcon";
-import { Tooltip } from "./basic/Tooltip/Tooltip";
-import { Dropzone } from "./basic/Dropzone/Dropzone";
 import { TemplateActionModal } from "./basic/TemplateActionModal/TemplateActionModal";
-import { ElvlButtonGroup } from "./basic/ElvlButtonGroup/ElvlButtonGroup";
+import { Toast } from "./basic/Toast/Toast";
+import { Tooltip } from "./basic/Tooltip/Tooltip";
+import { TypeBar } from "./basic/TypeBar/TypeBar";
 
 export {
   AttributeIcon,
@@ -42,11 +42,11 @@ export {
   Cloud,
   Checkbox,
   Dropzone,
-  ElvlButtonGroup,
   Header,
+  IconButtonGroup,
   Input,
-  MultiInput,
   Loader,
+  MultiInput,
   Panel,
   Tag,
   Table,
