+++ packages/client/src/pages/MainPage/containers/StatementEditorBox/StatementEditor/StatementEditorActantTable/StatementEditorActantTableRow.tsx
@@ -6,14 +6,11 @@ import {
   IResponseStatement,
   IStatementActant,
 } from "@shared/types";
-import {
-  AttributeIcon,
-  Button,
-  ButtonGroup,
-  ElvlButtonGroup,
-} from "components";
+import { excludedSuggesterEntities } from "Theme/constants";
+import { AttributeIcon, Button, ButtonGroup } from "components";
 import {
   AttributeButtonGroup,
+  ElvlButtonGroup,
   EntityDropzone,
   EntitySuggester,
   EntityTag,
@@ -26,15 +23,14 @@ import {
   useDrag,
   useDrop,
 } from "react-dnd";
-import { FaGripVertical, FaPlus, FaTrashAlt, FaUnlink } from "react-icons/fa";
+import { FaGripVertical, FaPlus, FaTrashAlt } from "react-icons/fa";
 import { UseMutationResult } from "react-query";
 import { setDraggedActantRow } from "redux/features/rowDnd/draggedActantRowSlice";
 import { useAppDispatch, useAppSelector } from "redux/hooks";
-import { excludedSuggesterEntities } from "Theme/constants";
 import {
+  DragItem,
   DraggedActantRowItem,
   DraggedPropRowCategory,
-  DragItem,
   FilteredActantObject,
   ItemTypes,
 } from "types";
