+++ packages/client/src/pages/MainPage/containers/StatementEditorBox/StatementEditor/StatementEditorActionTable/StatementEditorActionTableRow.tsx
@@ -1,13 +1,9 @@
 import { EntityEnums } from "@shared/enums";
 import { IProp, IResponseStatement } from "@shared/types";
 import { excludedSuggesterEntities } from "Theme/constants";
+import { AttributeIcon, Button, ButtonGroup } from "components";
 import {
-  AttributeIcon,
-  Button,
-  ButtonGroup,
   ElvlButtonGroup,
-} from "components";
-import {
   EntityDropzone,
   EntitySuggester,
   EntityTag,
